requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
python-multipart>=0.0.12
groq>=0.12.0
pandas>=2.2.3
openpyxl>=3.1.5
reportlab>=4.2.5
python-magic-bin==0.4.14
slowapi>=0.1.9
pydantic>=2.10.3
email-validator>=2.0.0
